pyproject.toml
[project]
name = "translator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
